Uploaded My Solution to Assignment
I uploaded my solution to the Pseudocode assignment which received a grade of 19.1 out of 20. The things that were marked wrong were minimum number of operations for Problem 2, maximum number of operations for Problem 2, maximum number of operations for Problem 3. The parts that got partial credit for were correct use of loop condition in Problem 4 (lost 0.2 out of 1.0 points) and maximum number of operations in Problem 4 (lost 0.1 out of 0.2 points).